// lib/types/supabase.ts
export interface definitions {
// ...
  comments_thread_with_user_vote: {
    id?: number;
    slug?: string;
    createdAt?: string;
    updatedAt?: string;
    title?: string;
    content?: string;
    isPublished?: boolean;
    authorId?: string;
    parentId?: number;
    live?: boolean;
    siteId?: number;
    isPinned?: boolean;
    isDeleted?: boolean;
    isApproved?: boolean;
    author?: string;
    votes?: number;
    upvotes?: number;
    downvotes?: number;
    depth?: number;
    path?: string;
    pathVotesRecent?: string;
    pathLeastRecent?: string;
    pathMostRecent?: string;
    userVoteValue?: number;
  };
// ...
}